requirements.txt
botbuilder-core==4.14.3
botbuilder-schema==4.14.3
botbuilder-connector==4.14.3
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6